Работа

+++ preprocessor/src/hdf5_processing.rs
@@ -8,26 +8,27 @@ use crate::types::MyConstData;
 pub fn write_data(data_path: &PathBuf, my_data: MyConstData) -> Result<()> {
     let file_path = data_path.join("my_daya.h5");
 
-    let file = File::create(file_path)?; // open for writing
-    let group = file.create_group("dir")?; // create a group
-    #[cfg(feature = "blosc")]
-    blosc_set_nthreads(2); // set number of blosc threads
-    let builder = group.new_dataset_builder();
-    #[cfg(feature = "blosc")]
-    let builder = builder.blosc_zstd(9, true); // zstd + shuffle
+    if file_path.exists() {
+        let file = File::open_rw(file_path)?;
 
-    // let my_pixels = TensorData::from(my_data.image.pixels);
-    // let my_keys = TensorData::from(my_data.keys_record.keys);
-    // let my_mouse = TensorData::from(my_data.keys_record.mouse);
+        let group = file.group("dir")?;
+        let ds = group.dataset("data")?;
+        ds.write(&[my_data])?;
+    } else {
+        let file = File::create(file_path)?; // open for writing
+        let group = file.create_group("dir")?; // create a group
+        #[cfg(feature = "blosc")]
+        blosc_set_nthreads(2); // set number of blosc threads
+        let builder = group.new_dataset_builder();
+        #[cfg(feature = "blosc")]
+        let builder = builder.blosc_zstd(9, true); // zstd + shuffle
 
-    let ds = builder
-        .with_data(&[my_data])
-        // finalize and write the dataset
-        .create("data")?;
-    // // create an attr with fixed shape but don't write the data
-    // let attr = ds.new_attr::<Color>().shape([3]).create("colors")?;
-    // // write the attr data
-    // attr.write(&[R, G, B])?;
+        let ds = builder
+            .with_data(&[my_data])
+            // finalize and write the dataset
+            .create("data")?;
+    }
+
     Ok(())
 }
 
@@ -39,14 +40,6 @@ pub fn read_data(data_path: &PathBuf) -> io::Result<Vec<MyConstData>> {
         let path = entry.path();
 
         if path.is_file() {
-            // let mut reader = csv::Reader::from_path(path).unwrap();
-            // for result in reader.deserialize() {
-            //     let record: CsvRecord = result.unwrap();
-            //     let keys_record = parse_csv_record(record);
-
-            //     dataset.push(keys_record);
-            // }
-
             let file = File::open(path)?; // open for reading
             let ds = file.dataset("dir/data")?; // open the dataset
